Writing day 3: methods
This basically finishes up the recognition system section.
I'm struggling to start on the synthesis system yet, as I feel like
I need to see the output of the recognition system in order to
start on the synthesis stuff. So I'm going to try to power through
and implement the stuff I just wrote about...