Transfer recognition network to Python package
From the `recog/auditory.ipynb` notebook. It was getting a bit too
big and hard to edit in there...

The `phd` package also contains a `doit` task for generating the
thesis from the LaTeX source in `paper`. It's installable from the
included `setup.py`.

Also, since I'm pretty much committed to using Brian's filters now,
I removed the Gammatone notebook. It's always in the history if
I need it!