Added experimentation with GVF
Although I followed the provided code closely, it doesn't seem to work.
There are no useful examples of the code working; it's designed to
be a plugin for other systems, so it's not straightforward
to compile and use.

So, a bit of a failure, but I think I understand the algorithm
better now.